requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
pytest
pytest-asyncio
httpx
pydantic-settings
pydantic[email]
bcrypt==4.0.1
passlib==1.7.4